[backend] Improve rate limit and add configuration for IP allow list (#15666)
