+++ docs/docs/deployment/configuration.md
@@ -73,6 +73,9 @@ Here are the configuration keys, for both containers (environment variables) and
 | app:graphql:armor_protection:block_field_suggestion   | APP__GRAPHQL__ARMOR_PROTECTION__BLOCK_FIELD_SUGGESTION | true                                                                                                                               | GraphQL armor protection, block the field suggestion                                                                |
 | app:notifier_authorized_functions                     | APP__NOTIFIER_AUTHORIZED_FUNCTIONS                     | see [`default.json`](https://github.com/OpenCTI-Platform/opencti/blob/master/opencti-platform/opencti-graphql/config/default.json) | Authorized function that can be used in webhook templates                                                           |
 | app:encryption_key                                    | APP__ENCRYPTION_KEY                                    | no default value                                                                                                                   | Encryption key used by platform cryptographic usage. Must be securely generated with `openssl rand -base64 32`      |
+| app:rate_protection:max_requests                    | APP__RATE_PROTECTION__MAX_REQUEST                      | 10000                                                                                                                              | Maximum of request allow per IP and per window duration, 0 will deny all                                            |
+| app:rate_protection:time_window                     | APP__RATE_PROTECTION__TIME_WINDOW                      | 1                                                                                                                                  | Time window for rate protection in seconds                                                                          |
+| app:rate_protection:ip_allow_list                   | APP__RATE_PROTECTION__IP_ALLOW_LIST                    | no default value                                                                                                                   | List of IP that are allowed to bypass the rate protection                                                           |
 
 #### Logging
